+++ include/sot/pattern-generator/exception-pg.h
