+++ include/sot/pattern-generator/next-step.h
