+++ include/sot/pattern-generator/pg.h
@@ -31,12 +31,12 @@
 
 /* SOT */
 
-#include <pinocchio/fwd.hpp>
 #include <dynamic-graph/entity.h>
 #include <dynamic-graph/pool.h>
 #include <dynamic-graph/signal-ptr.h>
 #include <dynamic-graph/signal-time-dependent.h>
-#include <sot-pattern-generator/exception-pg.h>
+#include <pinocchio/fwd.hpp>
+#include <sot/pattern-generator/exception-pg.h>
 #include <sot/core/flags.hh>
 #include <sot/core/matrix-geometry.hh>
