+++ include/sot/pattern-generator/selector.h
