+++ include/sot/pattern-generator/step-checker.h
