+++ src/exception-pg.cpp
@@ -17,7 +17,7 @@
  *
  * ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
 
-#include <sot-pattern-generator/exception-pg.h>
+#include <sot/pattern-generator/exception-pg.h>
 #include <stdarg.h>
 #include <cstdio>
