+++ src/step-computer-force.cpp
@@ -31,10 +31,10 @@
 
 #include <dynamic-graph/factory.h>
 #include <dynamic-graph/pool.h>
-#include <sot-pattern-generator/exception-pg.h>
-#include <sot-pattern-generator/step-checker.h>
-#include <sot-pattern-generator/step-computer-force.h>
-#include <sot-pattern-generator/step-queue.h>
+#include <sot/pattern-generator/exception-pg.h>
+#include <sot/pattern-generator/step-checker.h>
+#include <sot/pattern-generator/step-computer-force.h>
+#include <sot/pattern-generator/step-queue.h>
 #include <sot/core/debug.hh>
 #include <sot/core/macros-signal.hh>
 #include <sot/core/matrix-geometry.hh>
